RerouteTimeline refactored and tested
